// src/queries/projects/types.ts
export type IUpdateProject = {
  projectName: string;
  types: number[];
  status: number[];
  orientation: number[];
  architect: string;
  address: string;
  city: string;
  area: string;
  landSize: string;
  numberUnits: number;
  apartmentSize: string;
  numberFloors: number;
  numberParkings: number;
  handOverDate: string;
};